# improve index_select performance on CPU