--- a/tools/build_variables.bzl
+++ b/tools/build_variables.bzl
@@ -1133,6 +1133,7 @@ aten_native_source_codegen_list = [
     "aten/src/ATen/native/cpu/GridSamplerKernel.cpp",
     "aten/src/ATen/native/cpu/HistogramKernel.cpp",
     "aten/src/ATen/native/cpu/IndexKernel.cpp",
+    "aten/src/ATen/native/cpu/IndexSelectKernel.cpp",
     "aten/src/ATen/native/cpu/LerpKernel.cpp",
     "aten/src/ATen/native/cpu/LinearAlgebraKernel.cpp",
     "aten/src/ATen/native/cpu/MaxPooling.cpp",