@mingfeima 2022-07-07
> @mingfeima can you please rebase, as import/land attempt fail (build_variables.bzl were moved to top level folder)

Probably still need to update this one since we recently started the optimization project on PyG. Need to make sure the current optimization is able to cover the scenarios from PyG models...